Support an optional CTA link in a JS-created NotificationBar
This changeset adds support for an optional hyperlink added after the main text in a JS-triggered NotificationBar component.

An optional options object lets the developer specify the text for inside the link, the URL for its href and, optionally, another object of additional attributes which are all then applied to the link.

Includes tests.

Resolves https://github.com/mozilla/protocol/issues/460